Revert "Enhance DEVELOPMENT.md with new make commands for debugging database and API; update Makefile to include seed database command. Refactor main.go to improve SDK client initialization and login retry logic, and update database migration scripts to use BIGSERIAL and BIGINT for sensor IDs."
This reverts commit 1d80fad4288c8774f5817a6d8e2142271d9d1710.

+++ nexus-api/clients/database/schemas/postgres/migrations/20250404154934_sensor_moisture_data.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE sensor_moisture_data (
-    sensor_id BIGINT NOT NULL,
+    sensor_id INTEGER NOT NULL,
     id SERIAL PRIMARY KEY,
     date TIMESTAMP WITH TIME ZONE NOT NULL,
     soil_moisture DECIMAL NOT NULL,
