+++ nexus-api/clients/database/schemas/postgres/migrations/20250404154935_sensor_temperature_data.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE sensor_temperature_data (
     id SERIAL PRIMARY KEY,
-    sensor_id BIGINT NOT NULL,
+    sensor_id INTEGER NOT NULL,
     date TIMESTAMP WITH TIME ZONE NOT NULL,
     soil_temperature DECIMAL NOT NULL,
     FOREIGN KEY (sensor_id) REFERENCES sensors(id)
